+++ test_repro2.py
@@ -1,0 +1,13 @@
+import re
+_U = r"(?:req\.|request\.)"
+
+regex = re.compile(
+        r"(?:child_process\.exec|\bexecSync|\bexec|\bspawn|os\.system|os\.popen|subprocess\.(?:run|call|check_output|Popen))\s*\([^)]*"
+        + _U + r"|shell\s*=\s*True")
+
+text3 = """
+def run_cmd(cmd):
+  return os.popen(cmd).read()
+"""
+
+print("Regex 3:", regex.search(text3))