+++ test_repro4.py
@@ -1,0 +1,23 @@
+import re
+
+text = """
+class ImportPaste(graphene.Mutation):
+  result = graphene.String()
+
+  class Arguments:
+    host = graphene.String(required=True)
+    port = graphene.Int(required=False)
+    path = graphene.String(required=True)
+    scheme = graphene.String(required=True)
+
+  def mutate(self, info, host='pastebin.com', port=443, path='/', scheme="http"):
+    url = security.strip_dangerous_characters(f"{scheme}://{host}:{port}{path}")
+    cmd = helpers.run_cmd(f'curl --insecure {url}')
+"""
+
+# Let's see if there is any command injection pattern that can match this.
+regex = re.compile(
+        r"(?:child_process\.exec|\bexecSync|\bexec|\bspawn|os\.system|os\.popen|subprocess\.(?:run|call|check_output|Popen))\s*\([^)]*"
+        + r"(?:req\.|request\.)" + r"|shell\s*=\s*True")
+
+print("Regex 1:", regex.search(text))